+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 - Added `utils.listify` that will make sure a given value
   is always returned as list-like iterable.
   [gbastien]
+- Improved `system.get_git_tag` with new parameter to get last tag from all branches
+  [sgeulette]
 
 0.30 (2023-07-24)
 -----------------